refactor(ontology, provider): unify OCR types, provider re-exports from ontology
Remove duplicate Word/Line/Block/BlockKind/Page structs from the
provider crate. The provider now re-exports the identical types from
nvisy_ontology::artifacts::image. Drop the redundant Ocr prefix —
types are already namespaced by module.
