+++ crates/nvisy-ontology/src/artifacts/mod.rs
@@ -14,9 +14,7 @@ use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
 pub use self::audio::{AudioArtifacts, Transcription};
-pub use self::image::{
-    ImageArtifacts, OcrBlock, OcrBlockKind, OcrLine, OcrPage, OcrWord,
-};
+pub use self::image::{Block, BlockKind, ImageArtifacts, Line, Page, Word};
 pub use self::rich::RichArtifacts;
 pub use self::tabular::TabularArtifacts;
 pub use self::text::TextArtifacts;
