+++ crates/nvisy-provider/src/ocr/backend/output.rs
@@ -1,93 +1,22 @@
 //! OCR output types.
+//!
+//! The structural types ([`Word`], [`Line`], [`Block`], [`Page`]) are
+//! re-exported from [`nvisy_ontology::artifacts`]. [`ImageOutput`]
+//! wraps them with a [`ContentSource`] for provenance tracking.
 
 use nvisy_core::content::ContentSource;
-use nvisy_ontology::math::{BoundingBox, Polygon};
-use serde::{Deserialize, Serialize};
 
-/// A single word detected by OCR.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Word {
-    /// Recognised text content.
-    pub text: String,
-    /// Confidence score (0.0..=1.0), if the backend provides one.
-    pub confidence: Option<f64>,
-    /// Axis-aligned bounding box in pixel coordinates.
-    pub bbox: BoundingBox,
-    /// Polygon vertices for rotated or skewed text regions.
-    pub polygon: Option<Polygon>,
-}
-
-/// A line of text: ordered sequence of words.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Line {
-    /// Concatenated text from all words in this line.
-    pub text: String,
-    /// Line-level confidence, if the provider gives one.
-    pub confidence: Option<f64>,
-    /// Axis-aligned bounding box enclosing the line.
-    pub bbox: BoundingBox,
-    /// Polygon vertices for the line region.
-    pub polygon: Option<Polygon>,
-    /// Words in reading order.
-    pub words: Vec<Word>,
-}
-
-/// Classification of a block region.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum BlockKind {
-    /// Paragraph / prose.
-    Text,
-    /// Tabular content.
-    Table,
-    /// Figure / chart.
-    Figure,
-    /// Unclassified.
-    Other,
-}
-
-/// A block (paragraph, table cell, figure caption, etc.).
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Block {
-    /// Concatenated text from all lines in this block.
-    pub text: String,
-    /// Block-level confidence, if available.
-    pub confidence: Option<f64>,
-    /// Axis-aligned bounding box enclosing the block.
-    pub bbox: BoundingBox,
-    /// Polygon vertices for the block region.
-    pub polygon: Option<Polygon>,
-    /// Classification of this block.
-    pub kind: BlockKind,
-    /// Lines in reading order.
-    pub lines: Vec<Line>,
-}
-
-/// A single page of OCR results.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Page {
-    /// 1-based page number.
-    pub page_number: u32,
-    /// Page width in pixels, when known.
-    pub width: Option<f64>,
-    /// Page height in pixels, when known.
-    pub height: Option<f64>,
-    /// Blocks in reading order.
-    pub blocks: Vec<Block>,
-}
+// Re-export structural OCR types from ontology.
+pub use nvisy_ontology::artifacts::{Block, BlockKind, Line, Page, Word};
 
 /// Complete OCR output for one image/document.
 ///
 /// Groups detected text into a hierarchical tree of
 /// [`Page`] → [`Block`] → [`Line`] → [`Word`], together with a
 /// [`ContentSource`] derived from the input image for provenance tracking.
-///
-/// [`ContentSource`]: nvisy_core::content::ContentSource
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
 pub struct ImageOutput {
     /// Provenance: derived from the input's [`ContentSource`].
-    ///
-    /// [`ContentSource`]: nvisy_core::content::ContentSource
     pub source: ContentSource,
     /// Pages of OCR results.
     pub pages: Vec<Page>,
